feat: Implement a class that checks to see if you win the lotto
It was divided into two classes in total.
One managed the number of lottery winning numbers, lottery winning amounts,
and output messages using the enum class,
and the other created a class to compare the number of lotto bought by the user
with the winning number using the comparison function defined in the lotto class.